requirements.txt
# ============================================================
# ROOF TILE ORDERING SYSTEM — Requirements
# Last verified: April 2026
# ============================================================

# ------------------------------------------------------------
# CORE FRAMEWORK
# ------------------------------------------------------------
fastapi==0.136.0
uvicorn[standard]==0.44.0

# ------------------------------------------------------------
# WEBSOCKETS
# ------------------------------------------------------------
websockets==16.0

# ------------------------------------------------------------
# DATABASE
# ------------------------------------------------------------
sqlalchemy==2.0.49
aiosqlite==0.22.1

# ------------------------------------------------------------
# DATA VALIDATION
# ------------------------------------------------------------
pydantic==2.13.0
pydantic-settings==2.14.0

# ------------------------------------------------------------
# TEMPLATING
# ------------------------------------------------------------
jinja2==3.1.6
python-multipart==0.0.26

# ------------------------------------------------------------
# ENVIRONMENT
# ------------------------------------------------------------
python-dotenv==1.1.0

# ------------------------------------------------------------
# UTILITIES
# ------------------------------------------------------------
httpx==0.28.1
anyio==4.9.0
faker==37.1.0